Fixing report data types

+++ src/proxy.c
@@ -710,7 +710,7 @@ void proxy_addcell_in(struct proxy *p, struct cell *c, int type) {
   if (p->nr_cells_in == p->size_cells_in) {
 #ifdef SWIFT_DEBUG_CHECKS
     message("Growing proxy input cell list from %d to %d", p->size_cells_in,
-            p->size_cells_in * proxy_buffgrow);
+            (int)(p->size_cells_in * proxy_buffgrow));
 #endif /* SWIFT_DEBUG_CHECKS */
 
     p->size_cells_in *= proxy_buffgrow;
@@ -769,7 +769,7 @@ void proxy_addcell_out(struct proxy *p, struct cell *c, int type) {
   if (p->nr_cells_out == p->size_cells_out) {
 #ifdef SWIFT_DEBUG_CHECKS
     message("Growing proxy output cell list from %d to %d", p->size_cells_out,
-            p->size_cells_out * proxy_buffgrow);
+            (int)(p->size_cells_out * proxy_buffgrow));
 #endif /* SWIFT_DEBUG_CHECKS */
     p->size_cells_out *= proxy_buffgrow;
